CHROMIUM: Workaround arc++ sw_sync deadlock
Somehow if we pass the in-fence from surfaceflinger (a sw_sync) fence
down to the GPU driver and pass back a new GPU out-fence before the
in-fence is signaled we end up in a deadlock situation:

  * drm gpu_scheduler kthread is waiting on sw_sync in-fence to
    fire before queuing job on GPU
  * kms atomic helper is waiting on GPU out-fence to fire
  * the sw_sync in-fence never fires

This was exposed recently by 1e25f3b2825 ("android: Push in-fence-fd down to driver")
which replaced the synchronous sync_wait() with plumbing the in-fence
down to the driver.  Turns out the exact same thing also happens with
vulkan.

BUG=b:245687382, b:225664897, b:242815732
TEST=run aztec ruins vk onscreen (with USE=vulkan build), run dEQP-EGL

Change-Id: I1d5c8755e6d3d986ba1105bdca17596eb5d67c85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/mesa/+/4082615
Tested-by: Ikshwaku Chauhan <ikshwaku.chauhan@amd.corp-partner.google.com>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>