CHROMIUM: vulkan/runtime: Wait on Android acquire fence on CPU.
This works around stuck fences. Waiting on the CPU here can be a
"Quality of Implementation" issue and might result in lower perf.

However, this is overall very unlikely and this is the same behavior
as the Intel driver does so this is very low risk.

Ported for Mesa-22.0.3 form
https://chromium-review.googlesource.com/c/chromiumos/third_party/mesa/+/3158739/2

BUG=b:270499011,b:245687382,b:237232102,b:194858164,b:199406573,b:207433028
TEST=Pass CQ dry run

Change-Id: Ide418919f87ae61bf52e469f8b755844e2434520