gbm: Allow GBM allocation of YUV surfaces
YUV surfaces may potentially not be renderable by the GPU itself, but
they can be sampled from the GPU, and exported to other devices such as
video decoder which can render into them.

This is required by chrome/chromium v4l2 based video decode stack.  In
the context of ChromeOS, minigbm allows for allocating YUV gbm BOs.
This brings parity to minigbm, so that the v4l2 codec stack can also be
used on linux.

Suggested-by: https://gitlab.freedesktop.org/mesa/mesa/-/issues/7745
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/9079
Signed-off-by: Rob Clark <robdclark@chromium.org>